// src/common/api/entities/tutanota/TypeRefs.ts
export const ImportMailStateTypeRef: TypeRef<ImportMailState> = new TypeRef("tutanota", "ImportMailState")

export function createImportMailState(values: StrippedEntity<ImportMailState>): ImportMailState {
	return Object.assign(create(typeModels.ImportMailState, ImportMailStateTypeRef), values)
}

export type ImportMailState = {
	_type: TypeRef<ImportMailState>;

	_format: NumberString;
	_id: IdTuple;
	_ownerGroup: null | Id;
	_permissions: Id;
	failedMails: NumberString;
	status: NumberString;
	successfulMails: NumberString;
	totalMails: NumberString;

	importedMails: Id;
	targetFolder: IdTuple;
}
// ...
export type MailBag = {
	_type: TypeRef<MailBag>;

	_id: Id;

	mails: Id;
}
export const MailBoxTypeRef: TypeRef<MailBox> = new TypeRef("tutanota", "MailBox")

export function createMailBox(values: StrippedEntity<MailBox>): MailBox {
	return Object.assign(create(typeModels.MailBox, MailBoxTypeRef), values)
}

export type MailBox = {
	_type: TypeRef<MailBox>;
	_errors: Object;

	_format: NumberString;
	_id: Id;
	_ownerEncSessionKey: null | Uint8Array;
	_ownerGroup: null | Id;
	_ownerKeyVersion: null | NumberString;
	_permissions: Id;
	lastInfoDate: Date;

	archivedMailBags: MailBag[];
	currentMailBag: null | MailBag;
	folders: null | MailFolderRef;
	importedAttachments: Id;
	mailDetailsDrafts: null | MailDetailsDraftsRef;
	mailImportStates: Id;
	receivedAttachments: Id;
	sentAttachments: Id;
	spamResults: null | SpamResults;
}
// ...
export type MailDetailsDraftsRef = {
	_type: TypeRef<MailDetailsDraftsRef>;

	_id: Id;

	list: Id;
}
export const MailFolderTypeRef: TypeRef<MailFolder> = new TypeRef("tutanota", "MailFolder")

export function createMailFolder(values: StrippedEntity<MailFolder>): MailFolder {
	return Object.assign(create(typeModels.MailFolder, MailFolderTypeRef), values)
}

export type MailFolder = {
	_type: TypeRef<MailFolder>;
	_errors: Object;

	_format: NumberString;
	_id: IdTuple;
	_ownerEncSessionKey: null | Uint8Array;
	_ownerGroup: null | Id;
	_ownerKeyVersion: null | NumberString;
	_permissions: Id;
	color: null | string;
	folderType: NumberString;
	isMailSet: boolean;
	name: string;

	entries: Id;
	mails: Id;
	parentFolder: null | IdTuple;
}
export const MailFolderRefTypeRef: TypeRef<MailFolderRef> = new TypeRef("tutanota", "MailFolderRef")

export function createMailFolderRef(values: StrippedEntity<MailFolderRef>): MailFolderRef {
	return Object.assign(create(typeModels.MailFolderRef, MailFolderRefTypeRef), values)
}

export type MailFolderRef = {
	_type: TypeRef<MailFolderRef>;

	_id: Id;

	folders: Id;
}
// ...
export type SpamResults = {
	_type: TypeRef<SpamResults>;

	_id: Id;

	list: Id;
}